+++ forensics/src/output/encoder/helper/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod record;
+pub(crate) mod schema;
